Flatten a few more fixtures.
