+++ src/fixture/profile/color.rs
@@ -8,7 +8,8 @@ use number::{Phase, UnipolarFloat};
 
 use crate::master::FixtureGroupControls;
 
-use crate::fixture::prelude::*;use crate::osc::prelude::*;
+use crate::fixture::prelude::*;
+use crate::osc::prelude::*;
 use strum_macros::{Display as EnumDisplay, EnumIter, EnumString};
 
 #[derive(Default, Debug)]
@@ -251,4 +252,25 @@ impl AnimationTarget {
     pub fn is_unipolar(&self) -> bool {
         matches!(self, Self::Sat | Self::Val)
     }
+}
+
+impl Color {
+    pub fn map_controls(map: &mut GroupControlMap<ControlMessage>) {
+        map_color(map, &wrap_color);
+    }
+}
+
+impl HandleOscStateChange<StateChange> for Color {}
+
+fn wrap_color(sc: StateChange) -> ControlMessage {
+    sc
+}
+
+pub fn map_color<F, T>(map: &mut GroupControlMap<T>, wrap: &'static F)
+where
+    F: Fn(StateChange) -> T + 'static,
+{
+    map.add_phase("Hue", move |v| wrap(StateChange::Hue(v)));
+    map.add_unipolar("Sat", move |v| wrap(StateChange::Sat(v)));
+    map.add_unipolar("Val", move |v| wrap(StateChange::Val(v)));
 }
