+++ src/osc/profile/mod.rs
@@ -1,15 +1,12 @@
-mod color;
 mod colordynamic;
 mod comet;
 mod dimmer;
 mod faderboard;
-mod freedom_fries;
 pub mod generic;
 mod h2o;
 mod hypnotic;
 mod lumasphere;
 mod radiance;
-mod rotosphere_q3;
 mod rush_wizard;
 mod solar_system;
 mod starlight;
